Consolidates subroles into main role
Moves the apt repo setup tasks into a shared role.

Moves ossec-agent/defaults to main role.
The only default variable is for the networking iface,
which is used for resolving the IP address for registration.
This variable applies to both manager and agents, so it
makes sense to share the same defaults file.

Agents and manager can be configured differently via
group_vars, which is outside the scope of this role.

Moves agent and manager config tasks into main role.
Uses group membership for conditional includes
of task lists for agent and manager.